add location for admin service data

+++ run-scripts/run-election-manager.sh
@@ -9,6 +9,11 @@ cd "$(dirname "$0")"
 CONFIG=${VX_CONFIG_ROOT:-./config}
 source ${CONFIG}/read-vx-machine-config.sh
 
+if [ -z "${ADMIN_WORKSPACE:-}" ]; then
+  echo "error: please set ADMIN_WORKSPACE and try again" >&2
+  exit 1
+fi
+
 export PIPENV_VENV_IN_PROJECT=1
 export NODE_ENV=production
 (trap 'kill 0' SIGINT SIGHUP; make -C vxsuite/services/admin run & make -C vxsuite/services/smartcards run & make -C vxsuite/services/converter-ms-sems run & make -C vxsuite/frontends/election-manager run) | logger --tag votingworksapp
